chore: move PublicationCard to cards
